+++ RELEASES.md
@@ -2,6 +2,7 @@ Stock Additions 0.7.7
 ========================
  * LaneSpeed process is much more efficient, CPU usage down from 6% (causing lag alerts) to ~1%
  * Use numpy_fast.interp instead of np.interp for slight boost in performance
+ * Support white panda with experimental parameter. localizer performance may be reduced
 
 Version 0.7.7 (2020-07-20)
 ========================